==> proj2/notebooks/1.1-EDA.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import copy
warnings.filterwarnings('ignore')

# load Enron Emails data
docword_enron = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.enron.txt',
                            sep=" ", skiprows=3, header=None)
docword_enron.columns = ["docID", "wordID", "count"]
with open('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.enron.txt') as f:
    enron_d = int(f.readline().splitlines()[0])
    enron_w = int(f.readline().splitlines()[0])
    enron_nnz = int(f.readline().splitlines()[0])
vocab_enron = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/vocab.enron.txt', header=None)
vocab_enron.columns = ["word"]
# wordID starts from 1
vocab_enron.index += 1

# load NIPS full papers data
docword_nips = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.nips.txt',
                           sep=" ", skiprows=3, header=None)
docword_nips.columns = ["docID", "wordID", "count"]
with open('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.nips.txt') as f:
    nips_d = int(f.readline().splitlines()[0])
    nips_w = int(f.readline().splitlines()[0])
    nips_nnz = int(f.readline().splitlines()[0])
vocab_nips = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/vocab.nips.txt', header=None)
vocab_nips.columns = ["word"]
# wordID starts from 1
vocab_nips.index += 1

# load KOS blog entries data
docword_kos = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.kos.txt',
                          sep=" ", skiprows=3, header=None)
docword_kos.columns = ["docID", "wordID", "count"]
with open('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.kos.txt') as f:
    kos_d = int(f.readline().splitlines()[0])
    kos_w = int(f.readline().splitlines()[0])
    kos_nnz = int(f.readline().splitlines()[0])
vocab_kos = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/vocab.kos.txt', header=None)
vocab_kos.columns = ["word"]
# wordID starts from 1
vocab_kos.index += 1

# load NYTimes news articles data
docword_nytimes = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.nytimes.txt',
                          sep=" ", skiprows=3, header=None, nrows=655468)
docword_nytimes.columns = ["docID", "wordID", "count"]
with open('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.nytimes.txt') as f:
    nytimes_d = int(f.readline().splitlines()[0])
    nytimes_w = int(f.readline().splitlines()[0])
    nytimes_nnz = int(f.readline().splitlines()[0])
vocab_nytimes = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/vocab.nytimes.txt', header=None)
vocab_nytimes.columns = ["word"]
# wordID starts from 1
vocab_nytimes.index += 1

# load PubMed abstracts data
docword_pubmed = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.pubmed.txt',
                          sep=" ", skiprows=3, header=None, nrows=500018)
docword_pubmed.columns = ["docID", "wordID", "count"]
with open('/hdd/work/pw/4sem/ml/proj/proj2/data/docword.pubmed.txt') as f:
    pubmed_d = int(f.readline().splitlines()[0])
    pubmed_w = int(f.readline().splitlines()[0])
    pubmed_nnz = int(f.readline().splitlines()[0])
vocab_pubmed = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data/vocab.pubmed.txt',
                           header=None, sep='this string does not exist in a file')
vocab_pubmed.columns = ["word"]
# wordID starts from 1
vocab_pubmed.index += 1

docword_enron

docword_nips

docword_kos

docword_nytimes

docword_pubmed

docword_enron_cut = docword_enron.loc[docword_enron["docID"] <= 1500]
docword_nips_cut = docword_nips.loc[docword_nips["docID"] <= 1500]
docword_kos_cut = docword_kos.loc[docword_kos["docID"] <= 1500]
docword_nytimes_cut = docword_nytimes.loc[docword_nytimes["docID"] <= 1500]
docword_pubmed_cut = docword_pubmed.loc[docword_pubmed["docID"] <= 1500]

docword_enron_cut

vocab_enron

vocab_enron = vocab_enron.reset_index()

#df_temp = 'wordID':vocab_enron.index, 'word':vocab_enron.values})
#enron_temp = pd.merge(docword_enron_cut, df_temp, how='left', on='wordID')
#df_temp
vocab_enron = vocab_enron.reset_index()
data = [vocab_enron['index'], vocab_enron['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)
enron_temp = pd.merge(docword_enron_cut, df_temp, how='left', on='wordID')

df_temp

enron_temp

vocab_nips = vocab_nips.reset_index()
data = [vocab_nips['index'], vocab_nips['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)
nips_temp = pd.merge(docword_nips_cut, df_temp, how='left', on='wordID')

nips_temp

vocab_kos = vocab_kos.reset_index()
data = [vocab_kos['index'], vocab_kos['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)
kos_temp = pd.merge(docword_kos_cut, df_temp, how='left', on='wordID')

kos_temp

vocab_nytimes = vocab_nytimes.reset_index()
data = [vocab_nytimes['index'], vocab_nytimes['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)
nytimes_temp = pd.merge(docword_nytimes_cut, df_temp, how='left', on='wordID')

nytimes_temp

vocab_pubmed = vocab_pubmed.reset_index()
data = [vocab_pubmed['index'], vocab_pubmed['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)
pubmed_temp = pd.merge(docword_pubmed_cut, df_temp, how='left', on='wordID')

pubmed_temp

# enron_temp, nips_temp, kos_temp, nytimes_temp, pubmed_temp

nips_temp["docID"] += 1500

nips_temp

kos_temp["docID"] += 3000

kos_temp

nytimes_temp["docID"] += 4500

nytimes_temp

pubmed_temp["docID"] += 6000

pubmed_temp

df_merged = pd.concat([enron_temp, nips_temp, kos_temp, nytimes_temp, pubmed_temp])

df_merged = df_merged.reset_index()

df_merged

df_merged = df_merged.drop('index', 1)

df_merged

# vocab_enron, vocab_nips, vocab_kos, vocab_nytimes, vocab_pubmed

dict_merged = pd.concat([vocab_enron, vocab_nips, vocab_kos, vocab_nytimes, vocab_pubmed])

dict_merged = dict_merged.reset_index()

dict_merged

dict_merged = dict_merged.drop('level_0', 1)

dict_merged = dict_merged.drop('index', 1)

# df_merged, dict_merged

df_merged

dict_merged

dict_merged = dict_merged.reset_index()

data = [dict_merged['index'], dict_merged['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)

df_temp

df_merged = df_merged.drop('wordID', 1)

df_merged_temp = pd.merge(df_merged, df_temp, how='left', on='word')

df_merged_temp

df_merged

dict_merged

df_merged_testowo = dict_merged.drop_duplicates(subset='word', keep="first")

dict_merged

df_merged_testowo

df_merged_temp = pd.merge(df_merged, df_merged_testowo, how='left', on='word')

df_merged_temp

df_merged_temp.rename(columns = {'index':'wordID'}, inplace = True)

df_merged_temp

data_all = df_merged_temp

data_all

dict_all = df_merged_testowo

dict_all = dict_all.reset_index()

dict_all

dict_all = dict_all.drop('index', 1)

dict_all

data_all

data_all = data_all.drop('wordID', 1)

# data_all, dict_all

dict_all

data_all.to_csv('data_all.csv')
dict_all.to_csv('dict_all.csv')
