==> proj2/notebooks/2.0-preprocessing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import copy
warnings.filterwarnings('ignore')

data_all = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data2/data_all.csv')
dict_all = pd.read_csv('/hdd/work/pw/4sem/ml/proj/proj2/data2/dict_all.csv')

data_all

temp = dict_all[dict_all['word'].isin(data_all['word'])]

temp = temp.reset_index()

dictionary = temp.drop('index', 1)

data = data_all

# data, dictionary

dictionary

dictionary.to_csv('dict_before_spacy.csv')

! python -m spacy download en

import spacy
import pandas as pd
from tqdm.auto import tqdm 
import swifter
import plotly.express as px
from wordcloud import WordCloud
import setuptools
import textacy

dictionary

nlp = spacy.load("en_core_web_sm") # loading spacy model
dictionary['word_token']= dictionary['word'].astype(str)
dictionary['word_token'] = dictionary['word_token'].swifter.apply(nlp)

dictionary

dictionary

tokens = [token[0].lemma_ for token in dictionary['word_token'] if not token[0].is_stop if not token[0].is_punct]

tokens

tokens_unique = set(tokens)

len(tokens_unique)

tokens_unique

essunia = list(tokens_unique)

len(essunia)

data_all

data_all['word_token']= data_all['word'].astype(str)
data_all['word_token'] = data_all['word_token'].swifter.apply(nlp)
