==> proj2/notebooks/2.1-preprocessing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import copy
warnings.filterwarnings('ignore')

from google.colab import drive
drive.mount('/content/drive')

data_all = pd.read_csv('/content/drive/My Drive/datasets/data_all.csv')
dict_all = pd.read_csv('/content/drive/My Drive/datasets/dict_before_spacy.csv')
indeksY_pod = pd.read_csv('/content/drive/My Drive/datasets/indeksy_pod_idf.csv')

dict_all = dict_all.drop("Unnamed: 0", 1)
data_all = data_all.drop("Unnamed: 0", 1)

data_all

data = [dict_all.reset_index()['index'], dict_all['word']]
headers = ["wordID", "word"]
df_temp = pd.concat(data, axis = 1, keys = headers)
essa = pd.merge(data_all, df_temp, how='left', on='word')

essa = essa.drop("word", 1)

essa

indeksY_pod

lista_indeksow = indeksY_pod['Unnamed: 0'].to_list()

len(lista_indeksow)

essa_fixed = essa[essa['wordID'].isin(lista_indeksow)]

essa_fixed

document_lengths = essa_fixed.groupby(['docID', 'wordID'])['count'].count()

document_lengths

document_freqs = essa_fixed.groupby('wordID')['docID'].count()
document_freqs = pd.DataFrame(document_freqs)
document_freqs.rename(columns={'docID': 'df'}, inplace=True)
document_freqs

sum(document_freqs['df'])

document_freqs.describe()

dfs = np.array(document_freqs['df'])
dfs

idfs = np.log(7500/dfs)
document_freqs['idf'] = idfs
document_freqs

essa_testyn = essa.drop("word", 1)

essa

testyn = essa.pivot('docID', 'wordID', 'count')

testyn.info()

testyn = testyn.fillna(0)

testyn

document_freqs.to_csv('idf_of_all_words_fixed.csv')

testyn.to_csv('counters_all_words.csv')
