==> proj2/notebooks/3.0-modelling.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import copy
warnings.filterwarnings('ignore')

from google.colab import drive
drive.mount('/content/drive')

data = pd.read_csv('/content/drive/My Drive/WUM_PROJEKT2/complete_tf_idf.csv')
dictionary = pd.read_csv('/content/drive/My Drive/WUM_PROJEKT2/dictionary.csv')
dictionary = dictionary.drop("Unnamed: 0", 1)
data = data.drop("Unnamed: 0", 1)

data

data_fixed = data.drop(data.index[5223])

data_fixed

df1 = data_fixed[data_fixed.isna().any(axis=1)]
df1

tf_idfs = data_fixed.values
tf_idfs

print(np.any(np.isnan(tf_idfs)))
print(np.all(np.isfinite(tf_idfs)))
print(np.count_nonzero(np.isnan(tf_idfs)))
print(np.isnan(tf_idfs).sum())

# Perform the necessary imports
from sklearn.decomposition import PCA
#from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import make_pipeline
import matplotlib.pyplot as plt
from matplotlib.pyplot import figure

# Create scaler: scaler
#scaler = StandardScaler()

# Create a PCA instance: pca
pca = PCA()

# Create pipeline: pipeline
#pipeline = make_pipeline(pca)

# Fit the pipeline to 'samples'
pca.fit(tf_idfs)

# Plot the explained variances
features = range(pca.n_components_)
plt.figure(figsize=(10, 10))
plt.bar(features, pca.explained_variance_)
plt.xlabel('PCA feature')
plt.ylabel('variance')
plt.xticks(features)
plt.show()


pca.explained_variance_[1:4500]

tf_idfs

from sklearn.decomposition import TruncatedSVD
from sklearn.cluster import KMeans
from sklearn.pipeline import make_pipeline

svd = TruncatedSVD(n_components=4446)
kmeans = KMeans(n_clusters=5)
pipeline = make_pipeline(svd, kmeans)
pipeline.fit(tf_idfs)
labels = pipeline.predict(tf_idfs)

titles = range(7499)

# Create a DataFrame aligning labels and titles: df
df = pd.DataFrame({'label': labels, 'article': titles})

# Display df sorted by cluster label
print(df.sort_values('label'))

df.sort_values('label')[0:2000]

df['true_label'] = 0

df['true_label'][1500:3000] = 1
df['true_label'][3000:4500] = 2
df['true_label'][4500:6000] = 3
df['true_label'][6000:7499] = 4

df

crosstb = pd.crosstab(df['true_label'], df['label'])

crosstb
